+++ src/main/java/org/folio/InventoryKafkaTopic.java
@@ -19,6 +19,7 @@ public enum InventoryKafkaTopic implements KafkaTopic {
   LOCATION("location"),
   LIBRARY("library"),
   CAMPUS("campus"),
+  SUBJECT_TYPE("subject-types"),
   INSTITUTION("institution");
 
   private static final String DEFAULT_NUM_PARTITIONS_PROPERTY = "KAFKA_DOMAIN_TOPIC_NUM_PARTITIONS";
@@ -33,7 +34,8 @@ public enum InventoryKafkaTopic implements KafkaTopic {
     LOCATION, Pair.of("KAFKA_LOCATION_TOPIC_NUM_PARTITIONS", "1"),
     LIBRARY, Pair.of("KAFKA_LIBRARY_TOPIC_NUM_PARTITIONS", "1"),
     CAMPUS, Pair.of("KAFKA_CAMPUS_TOPIC_NUM_PARTITIONS", "1"),
-    INSTITUTION, Pair.of("KAFKA_INSTITUTION_TOPIC_NUM_PARTITIONS", "1")
+    INSTITUTION, Pair.of("KAFKA_INSTITUTION_TOPIC_NUM_PARTITIONS", "1"),
+    SUBJECT_TYPE, Pair.of("KAFKA_SUBJECT_TYPE_TOPIC_NUM_PARTITIONS", "1")
   );
 
   private final String topic;